feat(planification): phases en heures + Gantt aligné aux bornes de jour
Bornes de jour (Gantt + agenda) :
Les phases se convertissaient en events virtuels avec start_at à 8h00
et un end_at souvent NULL (champ end_date inexistant côté schéma).
Conséquence : la barre Gantt d'une phase de 1 j commençait au tiers
du jour de début et se terminait au tiers du jour suivant. On aligne
désormais sur 00:00 — la bande couvre exactement le bon nombre de
cases. La fin se calcule comme start + duration_days × 24 h, ce qui
permet aussi les durées sub-jour (0.5 = demi-journée jusqu'à midi).

Phases en heures :
- Backend : duration_days passe de Integer à Numeric(6,2). Migration
  idempotente dans init_db (ALTER COLUMN TYPE NUMERIC).
- Schémas Pydantic Phase{Create,Update,Read} : Optional[float].
- Formulaire de phase : 2 champs côte à côte « j » + « h » (heures
  0-7). Combinés en duration_days = j + h/8 (1 journée = 8 h). Une
  phase de 4 h = 0 j + 4 h = 0.5 jour stocké.

+++ backend/app/api/v1/endpoints/project_phases.py
@@ -40,7 +40,8 @@ class PhaseCreate(BaseModel):
     name: str = Field(..., min_length=1, max_length=255)
     position: int = Field(default=0, ge=0)
     start_date: Optional[date] = None
-    duration_days: Optional[int] = Field(default=None, ge=0, le=3650)
+    # Décimal pour exprimer des heures (ex. 0.5 = ½ journée = 4 h).
+    duration_days: Optional[float] = Field(default=None, ge=0, le=3650)
     notes: Optional[str] = None
     # Legacy scalar fields — toujours acceptés pour compat. Si les
     # listes ci-dessous sont fournies, elles priment.
@@ -54,7 +55,8 @@ class PhaseUpdate(BaseModel):
     name: Optional[str] = Field(default=None, min_length=1, max_length=255)
     position: Optional[int] = Field(default=None, ge=0)
     start_date: Optional[date] = None
-    duration_days: Optional[int] = Field(default=None, ge=0, le=3650)
+    # Décimal pour exprimer des heures (ex. 0.5 = ½ journée = 4 h).
+    duration_days: Optional[float] = Field(default=None, ge=0, le=3650)
     notes: Optional[str] = None
     assignee_employe_id: Optional[int] = None
     assignee_sous_traitant_id: Optional[int] = None
@@ -69,7 +71,7 @@ class PhaseRead(BaseModel):
     name: str
     position: int
     start_date: Optional[date]
-    duration_days: Optional[int]
+    duration_days: Optional[float]
     notes: Optional[str]
     # Champs scalaires legacy — renseignés au « primary » assignee
     # (= premier employé / sous-traitant de la liste) pour que les
